Changes to let dvm-dos-tem run in "spatial series" rather than "time series".
Notes from Fengming Yuan, 6/29/2012
---------------------------------------

Added 2 running subroutines in Runner.cpp, runmode2 and runmode3.
	- runmode2 is for timeseries,
	- runmode3 is spatial series.
The idea is that let the TEM run through a loop of 'cohort' (cell) at
 monthly time step. At the end of each time step, save all needed states
variables into the so-called 'restart' file. When starting the next time
step, the model will first read the initial conditions from that
'restart' file to continue with.

BUG: the restart file after each time-step is not output correctly.

There is one big changes in the code with regarding to the input data reading.
 Previously I let the 'cohort' (cell) ID to be searched throughout the input
.nc files each time – looks like that will cost a lot of time. Now the codes
in this aspect have changed so that in the beginning all 'cohort' (cell) ID
are read into a few vectors, and then data-reading can directly go into
the .nc files – no more searching (although it needs a little bit more memory).