Add script that can compile in debug mode w/ optimization off.
Not sure if this is the best way, but it works. Basically it
changes the Makefile to remove the optimization, then
compiles the model, and replaces the original Makefile.

The advantage here is that the original Makefile remains the default
so that if we get some tests going, the default behavior is to
compile normally and run normally - should prevent errors to do with
different results that may occur from the compiler optimization.