two bugs fixed, data clean-up simplified, and 'long' type avoided
(1) front-type in 'restart' bug fixed;
(2) Tobey caught a bug in 'climate' data processing for non-eq
runstages. It's fixed;
(3) data clean-up after one cohort is done previously is implemented
individually for data structure. It's too complicated and prone to
error. It can be done just by calling relevant constructor.
(4) When moving code from 32-bit to 64-bit, 'long' type definition could
cause issues. This is totally checked in this version of code. This
issue was found when comparing running DOS-TEM on my 32-bit windows XP
and on aeshna.