Changing paths in config files. Maybe not quite right yet?
At this point, the code has been updated from aeshna
and the calibration version is incorporated in its own
sub-directory of the main model. The main extrapolation
code and the calibration code will both compile on aeshna.

The calibration version is built with the tool "ant" and the
script "build.xml". The paths in this script are set up to
rely on the source code of the extrapolation version for building
the shared library.

Once built, the calibration version can be run with this command:

    $ java -Djava.library.path="lib" -jar dvm-dos-tem-calibrator.jar

The program will should then launch. If running on aeshna, remember
to ssh with the -X or -Y options so the display will work.

The code may compile on a Mac, but it has not been tested yet.