updated codes with most recently developments
the following critical developments:
(1) dead moss layer is separated as a horizon with different
decomposition rate (to be calibrated), which previously was same as
litter decomposition rate;
(2) environmental module (biophysical processes) has been rewritten for
soil thermal/hydrological processes. Previously there are yearly
oscillation of thermal properties, and this update is supposed to be
improve in this issue.

To be improving: (1) 'fpc' is not really effectively working currently
due to not-yet-identified issues; (2)  organic horizon layer division
needs to incorporate Fan's methods; and  (3) methane module
incorporation from Fan's work.