more major refactoring...COMPILES, RUNS
Concept now is that the CalController won't be a member of any
of the existing classes. Instead, there will be some sort of
calibrationON/OFF parameter that is passed down thru the call chain
(ArgHandler -> Runner -> RunCohort).

In RunCohort::run_timeseries the CalController will be instantiated
if in cal mode. Otherwise, it will be a null pointer so that there
will be minimal performance impact in extrapolation mode.

 - overhaul CalController class
 - pull all prototyping code out of TEM.cpp
 - remove CalController member from RunCohort
 - add conditional pointer initialization to RunCohort (hard coded
   condition right now)