Modify so that in calibration mode, the model pauses after the initial 100 year environmental module-only warm-up period.
In RunCohort, move cal controller pointer up one function call and then
 pass the pointer into subsequent functions. Will allows for control
of model by calibration controller from different places in the code.

In CalController, add a general "pause" function that posts a handler
to the io_service. This seems to be a way to run handlers programatically
instead of only running them when a signal is caught.