Changes the reload calparbgc feature of the calibration interface
so that it 'propogates' the new parameters all the way to the
Cohort's internal data. Prior to this only the CohortLookup
was updated. This meant that the user could edit and reload
the cmt_calparbgc.txt file and successfully reload it, but
the changes would not necessarily be reflected in the simulation.