Squash merging old 'tobeycarman/p1' into 'parallel1' branch for a fresh start.
Tobey's old 'p1' branch had a handful of junk commits and
we decided that we didn't need to retain the exact history
of the branch.

This code is definitely in rough draft form! The outputs
have not been checked at all and I am not actually sure
if the model in parallel mode is actually even producing
valid outputs. I recall that when I left off with this
development, I was working on getting the RestartData
structure to be reliably passed between timesteps and
run stages (eq, sp, tr, sc).

Below are some notes about what is happening with different
 ways of running the model:

Runs a single site thru all years. Default command line
options. One processor:

    $ ./DVMDOSTEM

Runs each cohort thru all timesteps before moving on to
the next cohort. One processor:

    $ ./DVMDOSTEM -m regnrun -f config/controlfile_regn.txt --space-time-config regner1
    Verification reuimentary - needs more programming help!
    Setting up logging...
    DVM-DOS-TEM-Test-06-cohorts: 6  cohorts to be run @eq
       from:  1  to:  6
    cohort: 1 - running!
    cohort: 2 - running!
    cohort: 3 - running!
    cohort: 4 - running!
    cohort: 5 - running!
    cohort: 6 - running!
    DONE.

Doesn't work in non-parallel mode to run over each timestep
before moving on to the next cohort. This could be fixed
pretty easily in the code I think:

    $ ./DVMDOSTEM -m regnrun -f config/controlfile_regn.txt --space-time-config regner2
    Verification reuimentary - needs more programming help!
    Setting up logging...
    DVM-DOS-TEM-Test-06-cohorts: 6  cohorts to be run @eq
       from:  1  to:  6
    You are processor 0 of 1 available on this system.
    This is a master/slave program and requires more than one processor to run when compiled with the WITHMPI flag.
    EXITING...

Use MPI to run in parallel. In this mode, each slave
processor will run each cohort in its list of assigned
cohorts for one year and then go back to the beginning
of its cohort list before advancing to the next year:

    $ mpirun -np 4 DVMDOSTEM -m regnrun -f config/controlfile_regn.txt --space-time-config regner2

I believe this is not functional yet. It runs, but I
think it is not sharing the work - simply running the
same code on however many processors you specify:

    $ mpirun -np 4 DVMDOSTEM -l note -m regnrun -f config/controlfile_regn.txt --space-time-config regner1




Conflicts:
	Makefile
	src/ArgHandler.cpp
	src/TEM.cpp
	src/assembler/RunCohort.cpp
	src/assembler/Runner.cpp
	src/assembler/Runner.h
	src/data/RestartData.cpp
	src/data/RestartData.h
	src/inc/timeconst.h