refactor calibration targets into their own file, add cmd line options...
for selecting which community type to display targets for (can select by
name or number).

For now communities should be hard-coded into the the top of the
calibration_targets.py file, but there is a rought start and writing
and excel file. This could be improved and extended to read an excel file.