Add a few named scopes for logging.
This basically just adds some convenient tags to log records coming
out of a certain scope. It should make the logs easier to grep thru.
For example with --log-level=debug, there is a huge volume of messages
coming out of some water check. To see the debug log w/o these messages
you could do:

    $ ./dvmdostem -l debug 2>&1 | grep -v h20

(Remember to redirect stderr to stdout for the pipe to grep to work!)
I expect we will need to add/remove/change these scopes, but this is a
start.