Fix bug with reading fire statistics, update firestatistics file.
There was an error with how the netcdf file was read, resulting in
the fires always occurring in the winter. We also fixed the firestatistics.nc
file to have more reasonable values.

We set PFSEASON to favor summer fires, and have essentially set the
fire severity to the second level for now. Still not completely
confident how these values are being handled/used in the code.

netcdf firestatistics {
dimensions:
	GFIREID = UNLIMITED ; // (1 currently)
	GFSEASONNO = 4 ;
	GFSIZENO = 5 ;
variables:
	int PFSEASON(GFIREID, GFSEASONNO) ;
	int PFSIZE(GFIREID, GFSIZENO) ;
	int FRI(GFIREID) ;
	int GFIREID(GFIREID) ;
		GFIREID:long_name = "GFIREID" ;
data:

 PFSEASON =
  0, 15, 85, 0 ;

 PFSIZE =
  0, 1, 0, 0, 0 ;

 FRI = 1000 ;

 GFIREID = 195 ;
}