Handle overflow in SoilLookup::deriveParam5TextureEnv(..).
Previous behavior ignored overflow and resulted in -inf value(s) in
the psi[][] array. This implementation avoids -inf and instead sets
values that would overflow a float to +/-FLT_MAX.

This issue was discovered by using #include <fenv.h> and
feenableexcept(..) to track down the location of the floating
point error that was setting -inf in the psi array.