Fix error with negative clouds.
Adds routine to force negative numbers to zero, then
uses routine to modify the girr and nirr daily containers so
that all negative values are set to zero. Also adds assertions to
Climate::create_clouds() to check that nirr and girr are >= 0.0.

More info:
A problem with the interpolation routine is leading to negative values
for nirr and girr (specifically if a monthly nirr or girr value is
zero and the preceeding month has a value > 0, then the interpolation
will overshoot slightly and we end up with slightly negative numbers
in the daily nirr or girr containers). These negative values in
in turn led to negative values for clouds in the daily cloud container.