refactor bootstrap.sh into two separate files, remove vagrant/ direcotory.
The main advantage to this is that we can have the cloned
repos and various dotfiles (.bashrc, .vimrc) be applied
to the "vagrant" user (instead of root) without needing
to change users in the bootstrap script. If we run everything
in one script, and don't change users, then the cloned repos
and all custom settings apply to the "root" user, and not
the default vagrant user.

An advantage of two files over, say one file with several
functions, is that we don't have to deal with formatting/
indenting the HEREDOCs used for the dotfiles.

Removing the vagrant directory should make using the shared
(with host) vagrant folders a little less confusing?