Change the way the annual values of the pools are computed:
Before - annual pools are computed by averaging the monthly values.
This is incorrect and make the checking of the carbon and nitrogen balances
 a posteriori (after running the model) impossible.

Now - annual pools are equal to the monthly value in December.
The carbon and nitrogen balance can now be checked after running the model.
For instance, the annual change in vegetation carbon pool
(i.e. the December value of year n minus the December value of year n-1)
should equal the NPP of year n minus LTRFALL of year n minus BURNVEGC of year n