Utility functions: make get_timeseries a template function, add template function get_scalar.
In the process of rebasing the fix-mineral-soil branch onto devel
and separating its changes from the fire branch, I ran into a few changes
on the mineral branch that depended on work in the fire branch.
Cherry picking the exact commits proved difficult, so I ended up
essentially building this commit by hand. Not sure what the implications
will be when merging the lines of development back together...