calibration-viewer.py: fix issue with plotting from archives, axes labels.
The calibration viewer was never updated to be able to read from an
archive created using the newer layout for calibration files.

Also change a setting so that the axes are never labeled using scientific
notation or the "offset" method (i.e. +5e6)

Now if you run the model and then create an archive from the calibration-json
files doing something like this:

    $ tar -zcvf initsoln-fix.tar.gz /tmp/dvmdostem/calibration/

the calibration viewer should be able to read and plot from the archive by
doing something like this:

    $ ../calibration-viewer.py \
           --from-archive ../initsoln-fix.tar.gz \
           --suite VegSoil \
           --tar-cmtnum 5 \
           --no-show \
           --pft 0 \
           --save-name initsoln-fix-VegSoil-pft0