Adjust initdmossc parameter(cmt_bgcsoil.txt)
This parameter needs match up with the initdmossthick in
cmt_dimground.txt.

Basically the amount of C in the dead moss layer should match up with
the inital dead moss thickness as defined by the relationship in
Yi et al 2009. From conversation with H. Genent, Spring 2016, we decided
that 10cm is a reasonable guess for starting moss thickness