Initialize moss layer C when initializing organic soil layer C.
Previously there was no way for moss layers to be assigned C.
This commit makes it so that moss layers are initialized with raw C
when the organic soil layers are initialized, (and in the same manner).

This seems to fix an issue with soil temperatures going to NaN.

We think the setup for the bug was as follows:
 - Once model is running, moss layer thickness is re-calculated
   based on C content
 - No C was assigned to the moss layer, so thickness was being set to zero.
 - It seems likely that calculating temperatre for a zero-thickness layer
   might result in a NaN.

By assigning C to the moss layers (rawc, soma, sompr, somcr), the thickness
of the moss layers does not immediately drop to zero. The layer report
shows a resonable temperature and thickness profile.

This seems to affect the dynamics in several ways:
 - the vegetation dynamics seem to have changed slightly, with
   more activity and some kind of cycle that alternates every other year
 - the water table and active layer depth seems to behave better