Fix issue with reported 'tcond'.
It appears that the Layer.tcond variable was never really used, so this
commit attempts to use it. This was discovered by watching the values
of tcond printed out in the Ground::layer_report_string and noticing that
they never strayed from zero. Looking at the code, it appears that the value
for tcond was always used as a local variable given up as a result of the
Layer::getThermalConductivity() function.

So this commit sets Layer.tcond in the getThermalConductivity() function.

It does not appear that this issue was a problem beyond making the Layer
report useless for this column.