Fix model dynamics issues by intializing everything to zero.
During the IO-refactor effort one of the changes was to explicitly
initalize all value of all structs using "flag" values like "-77777".
The idea was that these flags would make it easy to trace down locations
where certain values were not being set correctly. While this maybe be true
in some instances, initializing to these "flag" values also signifinctly
affected many model dynamics, for example shutting off the soil dynamics
and forcing Rh to zero. This commit initialzes everything to zero and
we can see much more dynamic behavior across all suites of the model.

It may be useful to initialize to "flag" values in the future, but for
now it seems that initializing everything to zero provides better results.