Comment out "dissolved organic Nitrogen drainage" code - improves Soil C closure.
With N feed ON and this dissolved N drainage concept in place, each summer,
delta C (using the caljsoncheckfunctions.py script) would drop to zero,
despite there being fluxes from the soil carbon pool. This erronious
zero-value for change in C pools prevented the Soil C cycle from closing
during summer months.

The problem appears to be related to the Integrator failing to converge in
conditions where "dondrain" exceeded the del_sois.orgn value, sending the
del_sois.orgn values for one or more layers negative. In this case, the
Integrator would fall back to using the "oldstate" for the timestep.

Commenting out this code, the soil C cycle now closes for the first 21
years of the run, and then starts failing to close on month 7 of year 22.