caljsoncheckfunctions.py: improve image plotting.
 - add refinements for colormaps.
 - add refinements for masking out zero and NaN values

All of this was terribly tedious to get it to behave the way
I wanted. Things to remember:
 - it seems to help to avoid rotating the legend text any
   thing other than 90 - makes it hard to be sure the labels
   are lining up in the right spots.
 - the masking data is kinda tricky. tried a bunch of things
   before finally getting the effect I was looking for.

The idea is to be able quickly see where the error is
significant. In other words, some places, where the error is
really really tiny, we can ignore it. I want my eyes to narrow
in on places and patterns whre the error is significantly non-zero.
"significant" in this case seems to be +/-1.0e-7 or so...
The way this code is setup, that might vary by machine.