re-work the --pid tag setup.
The previous arrangement was a little awkward. This still feels awkward,
but maybe is in an improvement?

Now the default behavior of dvmdostem is to put the calibration json data
in

    /tmp/dvmdostem/calibration

If you specify the --pid-tag then the data goes in:

    /tmp/dvmdostem-<PIDTAG>/calibration

You can also change the root of the tree for calibraiton json data
by modifying the appropriate section in the config/config.js file.

And I added a flag to the calibration-viewer.py so that you can specify
the "data path" (--data-path).

Still somewhat awkward, but seems to be working.