calibration-viewer.py - force blocking behavior.
not sure when this became a problem but at some point blocking seemed
to no longer be the default on my OSX machine. Not sure if this is a some
machine specific setting I made. But it doesn;t seem problematic to just
force this to be true and it prevents a few moments of confusiton when
you update from github and the plots no longer stay displayed. Seems like
it will be handy to leave this setting in the main repo.