Add sample data files that cover all years of input from SNAP's tifs.
These were created using the create_region_input.py script on Tobey's OSX
machine. I had to adjust a few small things in the create_region_input.py
script to get it to work that may not be comitted yet, but the script is
very close to working out-of-the-box. The fire files are still essnetially
placeholders, as we haven't finished the disturbance module yet.