Change to averaging climate over first 30 years (from 10 years).
The eq runstage uses a simplified climate that is averaged over the first
number of years of the transient (historic) data. For development we
had it set at averaging over a 10 year window. As we move closer to
"production" we want to average over a 30 year window.

It appears from Runner::run_years() (~line 67) that the Spinup stage is
already set to recycle the first 30 years of climate data.