Refactor TEM.cpp main() run stage code to use new RestartData scheme.
Diff is hard to read because much of the refactoring changed the
indentation, so while much of the code is the same, it shows up as modified.

Previously, there was some convoluted logic that checked for the
existence of the various restart files. The checks were implemented slightly
differently for each stage. Now with a static member function
added to RestartData, we can create the empty files before looping over
any of the run mask and then during the stage, we can be assured that the
files already exist, therfore obviating all the need to check for the files.

Also the current scheme comes much closer to standardizing the
the code for each stage.