Update parameters for CMT04 (Shrub Tundra) from E.Euskirchen hand-calibration.
In this parameterization, the organic N is still low, and the C and N
in leaves on the forbs & grasses are also too low (but the leaves are
low in the targets as well). Briefly tried lowering the cfall
and nfall leaf paramters for the grasses, but it didn't do much.