Change logging strategy for monthly and yearly checksumming.
Change things around so that it is possible to set all the checksumming
messages to the 'err' log level. Each checksumming function accumulates
a list of error messages, and then, if there are any errors, it sends each
item from the list to the log stream.

The code is kind of ugly, but here it seems to work.