Add feature to create_region_input.py that can lookup coordinates from lat/lon
Based on code from Unidata. Seems to be working pretty well, but
there seems to be some discrepancy between reported "closest match" values
and the lat/lon values for the corner pixels in datasets created with
the indices of the "closest match". Maybe something with center vs corners
of pixels?