modify diagnostics to allow for and new organization of output plots:
Now instead of showing the error balances for several pools side by side,
we show one panel for each pool, and each panel has error plot, deltas, and
the err/delta plot.

In addition the title of the figure is set and there is an command line
option (cmd line) to save the image.