Move parameter verification from Vegetation_Bgc to CohortLookup.
Now the parameter verifications includes a check that the cpart values
for leaf, stem and root all add up to 1, as well as a check that
the cpart values are correctly proportioned in relation to the
initvegc values. The new implementation also prints a formatted
table to the log that might be useful for diagnosing problems
with the input parameters.