calibration-viewer.py: change CMT lookup logic.
Previously the displayed target values were controlled by
the command line options. Now there is simply an argument
for whether or not to display the targets, but which targets
are shown is controlled by looking up the CMT code in the
firt available calibration-json file.

Remove --tar-cmtnum and --tar-cmtname implementation for
displaying target values on calibration plots.

Add implemantation that looks up the CMT number for displaying
target values in the first input file.