Change calibration-viewer to ignore interrupt signal (ctrl+c) and instead use Ctrl+g
This addresses a convenience issue for an the user while calibrating.
The problem was that the user might intend to pause the simulation,
but have the calibration viewer window in focus, so Ctrl+c would kill
the viewer instead of pausing the model. In the new implementation,
Ctrl+c is ignored, so the user must close the calibration-veiwer using
one of the following:
 - put the plot window in focus (active) and type ctrl+g
 - use the command line program kill to send signal 9 to the program
 - close the GUI window using the GUI controls.