Remove old parallel code and dependency on netcdfcpp.
This was originally written for the AIEM project and
used a Master/Slave structure to get control parallelism over grid cells
or cohorts as they were known pre-refactor. The code was written before
we undertook the major refactor and worked for the time-major loop
orientation. The driver code that actually used these Master and Slave
classes was removed during the refactor but fir future reference can be
found in commit: 310950381839, looking at TEM.cpp and the
Runner::regional_time_major(..) function.