KIND OF WORKING!!
the build setup is pretty bad, but I do believe it compiled and
created a test file in parallel on my Vagrant VM.

I am working off of GCC5.3.0 (from easybuild). Had to re-build jsoncpp
using GCC5.3.0, which created some issue with the ABI interfacing with
BOOST which was installed with yum (and compile with GCC 4.x presumably?)
So I am able to fix this by providing -D_GLIBCXX_USE_CXX11_ABI=0 to the
build of jsoncpp. The jsoncpp project's build system is kinda **cked, so
I am using easybuild to try and manage the process, but I haven't addressed
the install step yet, or the passing of the CXXFLAGS variable thru easy
build into the scons for jsonscpp yet. So I basically stopped the easybuild
before the compile step, and then changed into the easybuild directory where
it was running scons to compile jsoncpp. Then I exported the env var and
ran scons. And it worked. And then I staticly link to the jsoncpp object.
Gotta get this cleaned up...