Add time coord variable to climate inputs.
This is a little tricky - to keep CF conformant, need to use days since.
Really helpful to use Pandas to generate the series of datetime objects,
and then the netCDF4 helper function to generate the offsets that actually
get stored in the time coordinate variable. However tested this with ncview
and it worked beautifully - ncview is easily able to pickup and use the
time coordinate variable.