Adds basic plotting script for examining NetCDF outputs.
The script is very simple, and makes a few assumptions about the data.
It is hard coded to cell (0,0) and assumes that the only non-time
variable in the file specified is the proper variable name.
It does not yet explicitly allow for fronts, instead plotting the front
data with a few soil 'layers'.