Remove old (ar4) driving data, add "demo-data" directory with ar5 inputs.
Previously we maintained a number of sample datasets. Now we will only
maintain one sample dataset and other input datasets will be maintained
outside of this git repo. There is a script for keeping a local mirror up-
to-date with the main input catalog (on atlas server).

Update various files to reflect new path to demo data.