Change vagrant settings and boostrap script to work for an Ubuntu VM.
In the past we had tried using Fedora for the development VM, thinking it
may be closer to the "production" environment on atlas. It may be but I am
not convinced that it is worth the effort, because I find Ubuntu easier to
work with in general.

I was unable to get QTCreator to install, but I did find a python based
graphical debugger called gdbgui which is pretty nice, and easy to install.

As is, this VM does not have any effort toward making MPI or other
parallelism work.