Moves Vagrantfile and bootstrap scripts into subdirectory.
This will allow for maintaining different virtual machines. Vagrant has a
search path for Vagrantfiles, and there is no way to specify a specific
file to use so instead we have to keep the files in a separate directory.

To run Vagrant with the file in a subdirectory, you have 2 options:
 1) Set the VAGRANT_CWD environment variable
 2) Change into the sub directory and run the vargrant commands