Change comment in cmt_dimvegetation. WARNING!:
Will require update to PEcAn wrapper! The PEcAn wrapper for
dvmdostem relies on the comments in the parameter files to be able to map
parameter values to "priors" as they are called in PEcAn. This modification
will require updating the PEcAn code so that the initial LAI parameter
is found and appropriately modified by PEcAn.