Update VM specs and env setup scripts to work with lapacke.
Tested on dvmdostem-Inspiron-560, Ubuntu 16.04 VM,
and tobey's atlas account.

On Ubuntu (VM and dvmdostem-Inspiron-560), all that needs
to happen is to install the lapacke and openblas libraries
thru the system package manager.

On Tobey's atlas account which is a user account on CentOS,
the library is bundled with openblas which is provided thru
an EasyBuild module.