Add command line option to calibration-viewer.py to control blitting.
Recent builds on Ubuntu with stock packages for matplotlib, etc
are having problems with blitting in the animation so this option
allows us to turn off blitting at the command line.

See https://github.com/matplotlib/matplotlib/issues/11602
for more details.

If the performance is too slow without blitting we will have to
figure something else out. Tested this on a VM and it is acceptable.