Add code to read projected co2 file.
This should compile and run. When run the model will fail with an
exception if run with an old config file (i.e. one w/o a path to a
projected co2 file). It will also fail with an exception if there is no
file for the projected co2 data.