Another attempt at a quantified measure of the calibration and a way to print it.
This one computes both the rank as a % distance from the "truth" value
and then weights that by the amount of C that the parameter contributes
to the ecosystem. Not sure how to weight params that don't contribute C?

Relies on new functions in param_util.py that help with C contribution
calculation.

The print out is very crude here too. Needs to be a table or something more
readable.