Change char array declaration to be plenty long.
For some reason I only get this error on the PEcAn dev VM (Ubuntu 18.04,
using gcc 7.4.0), but there is a gcc warning about potentially overwriting
the buffer when using sprintf. Probably not an issue for us because we
know what the inputs to the string will be but simply making the initial
char array a little longer gets rid of the warning.