Fix segfault that occurs when checking water and running off the bottom of the stack.
For some reason this only shows up on the PEcAn VM (Ubuntu 18.04, gcc 7.x).
Previously, during the water check, as we iterated thru layers, we'd
assign the next layer (which could possibly be NULL if we are at the end
of the stack!) and then check for rock, breaking if there was rock.

Now we break the loop if there is rock encountered before trying to advance
the layer pointer.