calibration-viewer.py: add code allowing a path to a targets file to be provided on the command line.
Previously, the assumption was that the calibration_targets.py file lived
alongside the dvm-dos-tem/calibration/calibration-viewer.py script. This
only presents problems when the user wants to add a new community, but
the user is not ready to have the new community incorporated in the upstream
version control.

This new implementation allows for the user to specify a path to a calibration
targets file that will be used to display the target lines.