param_util.py: make --replace-cmt-block function safer:
instead of simply overwriting the original ("A") file, which could result
in accitendal data loss if this is used on files that are not under source
control, now the feature prints the new data to stdout where it can be
inspected and then redirected to a file if necessary. This is more in line
with the pattern used by the --enforce-initvegc option.