initial commit of input/ouput viewing tool.
had this working, maybe on a VM?, several months ago, but it has
been broken in other environments. I think that the enmvironment is
sensitive to the installation of gdal, geos, shapely, and geopandas.
On macOS "Shell is not a LinearRing" errors went away after I did
`pip unibstall shapely && pip install shapely --no-binary shapely`