Add custom config options to create_region_input.py script.
This allows for user to set completly custom paths for
where to find inputs. Also put in the paths for the GFDL-CM3
inputs on atlas. But because the GFDL-CM3 inputs are not
stored in the same directory tree as the ancillary files, it
ends up being easiest to use a totally custom config file
with all the paths explicitly set.