Test working with script and notebook file.
Idea is that we have have both in the repo. If a user
wants to see the diff, they look at the script version.
If they want to view the formatted notebook with outputs (i.e. online at
github or the nbviewer site) then they look at the .ipynb version.

The two files are kept in sync with a post save hook
in the jupyter_notebook_config.py file.