Test automatic script conversion.
Seems like it worked. The only code change was to add a period in the
comment. Note that the .py version changed a bit because it added
the comments at the top and converted ipython magics to
get_ipython().run_line_magic(...)

While initially confusing, this should not be an issue for future work.