Fix inline doctests in Sensitivity.py, fix bug when running...
... both inline and external test files when Sensitivity.py is run by
itself. In otherwords everything works if you run just

    $ python -m doctests_Sensitivity.py

or just the inline doctests:

    $ python -m doctest Sensitivity.py

But if you were to run like this:

    $ python Sensitivity.py

which gets both the inline and external test file tests, then
it failes with an error in one of the tests in doctests_Sensitivity.py.
No idea why...?

So this commit simply comments out the code that runs the the external
test file from within the include guard in Sensitivity.py.