add mechanism for dealing with calibration targets
The ideal workflow involves using a csv file with
both the parameters in it as well as the calibration
targets. The calibration targets are stored in a
differnent file and using a differnet format (python
object). There was not an existing way to store
comments, units, etc with the targets. So this adds
an additional field to the calibration targets.

In the long run, it might be better to have a meta dict
added to each CMT in the targets file. But that seemed
like too much typing for now - perhaps will auto-
generate later?