Add metadata tranlation for new columns...
This adds the detection and translation of the new columns (units,
description, comments, refs) to the conversion functionality.

This modifies the "CMT data dictionary" structure that is used elsewhere
in param_utils.py. More fields are added to the dictionary, one for each
column, using the format for example: "units_cmax", with the parameter
variable name prefixed with the column name and an underscore. Hopefully
this will not cause problems in other parts of the code that parse this
structure.

Also fixes the format of a few variable names in CMT00 in the fixed
width text format. Makes things easier to parse. Plan to use these tools
to propogate this change to the other CMTs.