Fix bug with the build image.
Accidentally had committed a work in progress state that took some
shortcuts for speeding up the builds when working on the Dockerfiles.

With this implementation, we copy the entire .git directory into the
build image. This is not terribly efficient, as the only reason we need
the .git directory is so that we can run git describe in the Makefile to
set the git SHA in the compiled binary. There is probably a cleaner more
efficient way to achieve this same result.