Better handle header comments in CMT data blocks.
THis makes the comment header of the FWT datablock be multiline (to
accomodate various metadata that can be in the csv files).

This will create problems with other functions that assume that the
header of the FWT CMT data block is only 1 or two lines. To be addressed
in future commits.