fix bug with gdal_translate call.
Not sure what happened - maybe due to a GDAL version bump.
Previously we had to 1) convert to netcdf and 2) subset

As of Nov 2022, the existing implementation throws an
inscrutable error about computing negative width or height.
Oddly it works for most of the files, but fails on the
climate files.

The new implementation in this commit does the conversion
and subsetting in one step. This is actually a bit more efficient
as we no longer need one of the temp files. And it seems to
work, at least with GDAL from conda and pip:
libgdal, 3.6.0, hf51d2d0_5, conda-forge
gdal, 3.6.0, py38h721c686_5, conda-forge
GDAL @ file:///home/conda/feedstock_root/build_artifacts/gdal-split_1668799523615/work/pybuild_h721c686/swig/python