Change Sensitivity.py so that it can handle multi-pft experiments.
Previously, the PFT specified for the parameters had to be consistent
across all the parameters. I.e. if the user wanted to analyze cmax, and
nmax they had to choose a single PFT that would be modified for each
parameter. This was enforced in the pftnum() function and printed out in
the info function.

Now the user can specify any number of pfts for each parameter, and the
parameters may use different PFTs. In the sample matrix columns, the
pftnumber is appended to the parameter name, like so: cmax_pft3