Re-arrange code to better support the multi-pft concept and aid in testability.
General idea is to differentiate between the "initial parameter run
directory" vs the "seedpath". The seedpath is the source where the
initial parameter values come from. The run directory is where the
initial parameters are copied to in order to setup the run (previously
called the "initial value run".

So it is possible to have a seedpath, but not yet have setup the
rundirectory. However in order to setup the run directories you must
have called set_working_directory() and set_seed_path().

Changes:
- move code that sets up list of parameter specificaitons into
function.
- change name from "initial_value_run" to "intial_params_rundir" to be
more explicit
 - change (simplify) tests in constructor