Normalize line endings to LF (Unix).
For some reason these files were left as DOS (CRLF) line endings. It has
been this way for years, and generally hasn't seemed to create problems
except that diffs showing ^M are confusing. And it seems better to have
the whole repo standardized. It is possible that this is a relic of
trying to be able to more eaily diff with DOS-TEM or older/original
files from DVM-DOS-TEM.