fix behavior for bad file path.
Previously the page would load but be blank if there was a bad file path
(hard coded path on startup, so hard to account for running on different
machines).

Now if there are no approprate files found in the provided folder, then
we return an empty geodata frame. And the maps still loads. This allows
us to remove hard-coded paths from the program and still present the
user a map and a input box where they can enter their path.

If the user enters a bad path, the previous view is re-loaded.