refactor the output_utils.py and example_experiment_0
This is a relatively large refactor that aims to consolidate code into
the output_utils.py script (rather than keeping it in the documentation
files). This will help for re-usability as well as management.
Previously, the example_experiment_0.rst file had a rather large
embedded code snippet with functions to open and load netcdf files. This
commit moves that logic to output_utils.py (and changes it a little as
well).

Details:
 - add dimensions, datetime index to return of stitch_stages() function
 - replace "build_full_dataframe" with "load_trsc_dataframe" function
   - the new func uses the existing stage stitching code (with new
     new return values noted above)
 - remove unneeded funcs (load_trsc, build_full_dataframe)

- update example_experiment_0.rst to use the new functions in
output_utils.py