Moves handling of plural header names, adds tests
This change should prevent the problem of a user
running --show-options then using those exact
values (which are the column headers from
output_spec) as parameters to --on which will
throw errors for both Compartments and Layers.

This also adds very basic testing for both new
CL flags to the doctest file.